# docs/api/openapi.yaml
openapi: 3.2.0
...
info:
  title: Imperio Industrial — API pública
  version: 1.5.0
  summary: API REST del MMO de simulación económica, industrial y logística "Imperio Industrial".
  description: |
    API pública del servidor autoritativo de **Imperio Industrial** (GDD/SAD v1.2).

    ## Principios del contrato de API

    - **Una sola API para humanos y bots** (igualdad de API literal, ADR-010): mismos
      endpoints y mismos rate limits lógicos. El motor no distingue el origen de un comando.
    - **Servidor autoritativo**: el cliente solo envía intenciones; toda validación
      (espacio, fondos, stock, plazos) ocurre server-side.
    - **REST para operaciones no urgentes**; los eventos en tiempo real del área de interés
      (movimiento de vehículos, alertas configuradas) se distribuyen por **WebSocket**
      a través del Notification/Event Gateway y quedan fuera de este documento.
    - **El tablón global es pull, no push**: se consulta bajo demanda con filtros
      (producto, ubicación, precio, plazo); nunca hay push mundial de mercado.
    - **Sim-time en el contrato**: todo plazo de dominio (contratos, producción, viajes)
      se define y transmite en sim-time (segundos desde el génesis del mundo, ratio 24×).
      La traducción a tiempo real es responsabilidad exclusiva de la UI. Las dos únicas
      mecánicas en tiempo real son la ventana de sorteo y el cooldown anti-parpadeo.
...
      (invariante del ledger).
...
      distinguibles.
...

    ## Estructura de respuestas

...

...
servers:
...
security:
  - sessionToken: []
...
paths:
...
  /world/buildings/{buildingId}/production-batches:
    parameters:
      - name: buildingId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/BuildingId'
    get:
      operationId: listProductionBatches
      tags: [world, production]
      summary: Cola de producción de un edificio
      description: El progreso del lote en curso es analítico (derivado bajo demanda), no por tick.
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/BatchStatus'
        - $ref: '#/components/parameters/cursor'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Lotes encolados/en curso.
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProductionBatch'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/Maintenance' }
    post:
      operationId: queueProductionBatches
      tags: [world, production]
      summary: Encolar lotes de producción
      description: |
        Encola uno o varios lotes de una receta soportada por el edificio. La producción
        pausa sin combustible (`paused_no_fuel`) o con salarios impagados
        (`paused_no_workers`) — cascada de insolvencia sin deuda (GDD 5.9).
      parameters:
        - $ref: '#/components/parameters/idempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductionBatchCreate'
      responses:
        '201':
          description: Lotes encolados.
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    $ref: '#/components/schemas/ProductionBatch'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422':
          description: 'Insumos insuficientes, receta no soportada, o edificio no operativo.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/Maintenance' }
  /world/production-batches/{batchId}:
    delete:
      operationId: cancelProductionBatch
      tags: [world, production]
      summary: Cancelar un lote encolado
      description: Cancela los lotes aún no producidos de la orden. Lo ya producido queda asentado.
      parameters:
        - name: batchId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/BatchId'
        - $ref: '#/components/parameters/idempotencyKey'
      responses:
        '200':
          description: Orden cancelada.
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    $ref: '#/components/schemas/ProductionBatch'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: La orden ya está completada o cancelada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/Maintenance' }

  # ───────────────────────── WORLD: FLOTA Y CARGAMENTOS ─────────────────────────
...
components:
  securitySchemes:
    sessionToken:
      type: http
      scheme: bearer
      description: |
        Token de sesión emitido por `POST /auth/sessions`. El servidor solo almacena su
        hash. Humanos y bots usan el mismo esquema (credenciales por cuenta); los bots
        acceden por un camino de red interno multiplexado con la misma superficie de API.
  parameters:
    idempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        Clave de idempotencia para **reintentos seguros de comandos que mueven
        valor**: si el cliente reintenta con la misma clave, el servidor reproduce
        la misma respuesta del primer intento — nunca hay doble ejecución.
      schema:
        type: string
        format: uuid
    cursor:
      name: cursor
      in: query
      description: Cursor opaco de paginación devuelto en `meta.next_cursor`.
      schema:
        type: string
    limit:
      name: limit
      in: query
      description: Tamaño máximo de página.
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
  responses:
    BadRequest:
      description: Comando malformado o unidades inválidas.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: Sesión ausente o expirada.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: Comandar recursos de otra corporación (autorización por propiedad), o vehículo SELLADO durante handoff.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: Entidad inexistente (UUID no resuelto).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    TooManyRequests:
      description: Rate limit excedido — **idéntico para humanos y bots** (decisión de balance, no solo de protección).
      headers:
        Retry-After:
          description: Segundos hasta poder reintentar.
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    InternalError:
      description: Error inesperado. Nunca deja garantías a medio bloquear — toda operación sobre dinero/stock es atómica en el ledger (todo-o-nada).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Maintenance:
      description: |
        Servicio temporalmente no disponible; SIEMPRE reintentable con la MISMA petición.
        Dos causas, distinguibles por el `code`:

        - `MAINTENANCE_WINDOW`: ventana de mantenimiento diaria (sim-time congelado de forma coordinada).
        - `SERIALIZATION_CONFLICT`: la operación chocó con otras concurrentes sobre las mismas
          cuentas del ledger y agotó su presupuesto de reintentos en el servidor. La transacción se
          revirtió ENTERA —no queda ningún efecto parcial, ni garantías a medio bloquear— así que
          reenviar la misma petición tras `Retry-After` es seguro y es lo esperado del cliente.
      headers:
        Retry-After:
          description: Segundos estimados hasta poder reintentar (fin de la ventana, o drenaje de la ráfaga de escrituras).
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            maintenance:
              value:
                error:
                  code: MAINTENANCE_WINDOW
                  message: El mundo está en su ventana de mantenimiento diaria; el sim-time está congelado
                  details: { retry_after_seconds: 900 }
            serialization_conflict:
              value:
                error:
                  code: SERIALIZATION_CONFLICT
                  message: la operación chocó con otras concurrentes y no se aplicó; reintenta la misma petición
                  details: { retry_after_seconds: 1 }
  schemas:
...
    SimTime:
      type: integer
      format: int64
      minimum: 0
      description: Segundos de sim-time desde el génesis del mundo (ratio 24× — 1 día de juego = 1 hora real). Todo plazo de dominio se expresa en sim-time.
      examples: [31104000]
...
    RecipeId:
      type: string
      format: uuid
      description: UUIDv7 de receta.
...
    BuildingId:
      type: string
      format: uuid
      description: UUIDv7 de edificio.
    BatchId:
      type: string
      format: uuid
      description: UUIDv7 de orden de producción.
...
    BatchStatus:
      type: string
      enum: [queued, running, paused_no_fuel, paused_no_workers, completed, cancelled]
...
    Meta:
      type: object
      description: Metadatos comunes a toda respuesta exitosa.
      required: [sim_time, server_time]
      properties:
        sim_time:
          type: string
          pattern: '^[0-9]{1,4}-[0-9]{3}-[0-9]{2}:[0-9]{2}$'
          description: 'Sim-time actual del mundo en formato legible `AÑO-DÍA-HH:MM` (p. ej. año 360 de juego, día 45).'
          examples: ["360-045-12:30"]
        sim_time_seconds:
          $ref: '#/components/schemas/SimTime'
          description: Sim-time actual en segundos desde el génesis (forma canónica para clientes).
        server_time:
          type: string
          format: date-time
          description: Wall-clock del servidor (solo informativo — ninguna regla de juego depende de él).
        next_cursor:
          type: string
          description: Cursor de la página siguiente en respuestas paginadas; ausente si no hay más resultados.
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: |
            Código estable de error de dominio. Valores documentados:
            `INSUFFICIENT_COLLATERAL`, `INSUFFICIENT_FUNDS`, `PUBLICATION_EXHAUSTED`,
            `CANCEL_COOLDOWN_ACTIVE`, `BELOW_MIN_LOT`, `STOCK_ALREADY_RESERVED`,
            `PLACEMENT_INVALID`, `NOT_RESOURCE_OWNER`, `VEHICLE_SEALED`,
            `NO_ROUTE_FOUND`, `MAINTENANCE_WINDOW`, `SERIALIZATION_CONFLICT`,
            `RATE_LIMITED`, `VALIDATION_ERROR`, `NOT_FOUND`, `UNAUTHORIZED`,
            `INTERNAL`.
          examples: [INSUFFICIENT_COLLATERAL]
        message:
          type: string
          description: Descripción legible del error.
        details:
          type: object
          additionalProperties: true
          description: Contexto estructurado del error (importes como strings de punto fijo).
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/Error'
      examples:
        - error:
            code: INSUFFICIENT_COLLATERAL
            message: La garantía disponible no cubre la publicación solicitada
            details: { required: "1000", available: "740" }

    # ── Auth ──
...
    ProductionBatchCreate:
      type: object
      required: [recipe_id, batches_queued]
      properties:
        recipe_id:
          $ref: '#/components/schemas/RecipeId'
        batches_queued:
          type: integer
          minimum: 1
        queue_position:
          type: integer
          minimum: 0
          description: Posición deseada en la cola (por defecto, al final).
    ProductionBatch:
      type: object
      required: [id, building_id, recipe_id, batches_queued, batches_done, status, queue_position]
      properties:
        id:
          $ref: '#/components/schemas/BatchId'
        building_id:
          $ref: '#/components/schemas/BuildingId'
        recipe_id:
          $ref: '#/components/schemas/RecipeId'
        batches_queued:
          type: integer
          minimum: 1
        batches_done:
          type: integer
          minimum: 0
        status:
          $ref: '#/components/schemas/BatchStatus'
        queue_position:
          type: integer
          minimum: 0
        started_at_sim:
          $ref: '#/components/schemas/SimTime'
        progress_pct:
          type: number
          minimum: 0
          maximum: 100
          description: Progreso del lote en curso, **derivado analíticamente** en el momento de la consulta (no persiste).
        eta_sim:
          $ref: '#/components/schemas/SimTime'
          description: Sim-time estimado de finalización del lote en curso.

    # ── World: flota y cargamentos ──
...